Safeguard first account creation
After we used sargon methods to create a new account, it seems that the profile state change is received quicker in MainViewModel than the CreateAccountViewModel can react and send the event to navigate to confirmation screen. This commit aims to lock AppState from changing when receiving a Profile when the first account creation is still in progress.